Mobile: Toolbar color depends on sidebar
Toolbar should reflect the colors of the current view, whether it's a page or the sidebar.

+++ src/ios.m
@@ -327,7 +327,7 @@ iBool isPhone_iOS(void) {
 }
 
 int displayRefreshRate_iOS(void) {
-    return uiWindow_(get_Window()).screen.maximumFramesPerSecond;
+    return (int) uiWindow_(get_Window()).screen.maximumFramesPerSecond;
 }
 
 void setupWindow_iOS(iWindow *window) {
